LayoutTests/http/tests/websocket/tests/hixie76/handshake-fail-by-sub-protocol-mismatch.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../../js-test-resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script type="text/javascript">
description("Test WebSocket handshake fail if sub protocol name mismatches.");

window.jsTestIsAsync = true;
if (window.testRunner)
    testRunner.overridePreference("WebKitHixie76WebSocketProtocolEnabled", 1);

var protocol;

function endTest()
{
    shouldBeUndefined("protocol");
    clearTimeout(timeoutID);
    finishJSTest();
}

var url = "ws://localhost:8880/websocket/tests/hixie76/protocol-test?protocol=available-protocol";
var ws = new WebSocket(url, "requestting-protocol");

ws.onopen = function()
{
    debug("Connected");
};

ws.onmessage = function (messageEvent)
{
    protocol = messageEvent.data;
    ws.close();
};

ws.onclose = function()
{
    endTest();
};

function timeOutCallback()
{
    debug("Timed out in state: " + ws.readyState);
    endTest();
}

var timeoutID = setTimeout(timeOutCallback, 3000);

</script>
<script src="../../../../js-test-resources/js-test-post.js"></script>
</body>
</html>
